+++ src/curibio/sdk/__init__.py
@@ -16,6 +16,7 @@ from .constants import CHART_WINDOW_NUM_SECONDS
 from .constants import CONTINUOUS_WAVEFORM_SHEET_NAME
 from .constants import DEFAULT_CELL_WIDTH
 from .constants import EXCEL_OPTICAL_METADATA_CELLS
+from .constants import FORCE_FREQUENCY_RELATIONSHIP_SHEET
 from .constants import FULL_CHART_SHEET_NAME
 from .constants import INTERPOLATED_DATA_PERIOD_CMS
 from .constants import INTERPOLATED_DATA_PERIOD_SECONDS
@@ -33,6 +34,7 @@ from .constants import PER_TWITCH_METRICS_SHEET_NAME
 from .constants import SECONDS_PER_CELL
 from .constants import SNAPSHOT_CHART_SHEET_NAME
 from .constants import TSP_TO_DEFAULT_FILTER_UUID
+from .constants import TWITCH_FREQUENCIES_CHART_SHEET_NAME
 from .constants import TWITCHES_POINT_UP_UUID
 from .excel_well_file import ExcelWellFile
 from .exceptions import MetadataNotFoundError
@@ -75,4 +77,6 @@ __all__ = [
     "SECONDS_PER_CELL",
     "PER_TWITCH_METRICS_SHEET_NAME",
     "NUMBER_OF_PER_TWITCH_METRICS",
+    "TWITCH_FREQUENCIES_CHART_SHEET_NAME",
+    "FORCE_FREQUENCY_RELATIONSHIP_SHEET",
 ]