+++ tests/test_plate_recording.py
@@ -5,6 +5,9 @@ To create a file to look at: python3 -c "import os; from curibio.sdk import Plat
 To create a file to look at: python3 -c "import os; from curibio.sdk import PlateRecording; PlateRecording([os.path.join('tests','h5','v0.3.1','MA201110001__2020_09_03_213024__A3.h5')]).write_xlsx('.',file_name='temp.xlsx')"
 To create a file to look at: python3 -c "import os; from curibio.sdk import PlateRecording; PlateRecording.from_directory(os.path.join('tests','h5','v0.3.1')).write_xlsx('.',file_name='temp.xlsx')"
 
+python3 -c "import os; from curibio.sdk import PlateRecording; PlateRecording([os.path.join('tests','h5','v0.3.1','MA201110001__2020_09_03_213024','MA201110001__2020_09_03_213024__A1.h5',),os.path.join('tests','h5','v0.3.1','MA201110001__2020_09_03_213024','MA201110001__2020_09_03_213024__B2.h5',),]).write_xlsx('.',file_name='temp.xlsx')"
+
+
 """
 import datetime
 import os
@@ -17,6 +20,7 @@ from curibio.sdk import __version__
 from curibio.sdk import AGGREGATE_METRICS_SHEET_NAME
 from curibio.sdk import CALCULATED_METRIC_DISPLAY_NAMES
 from curibio.sdk import CONTINUOUS_WAVEFORM_SHEET_NAME
+from curibio.sdk import FORCE_FREQUENCY_RELATIONSHIP_SHEET
 from curibio.sdk import INTERPOLATED_DATA_PERIOD_CMS
 from curibio.sdk import METADATA_EXCEL_SHEET_NAME
 from curibio.sdk import METADATA_INSTRUMENT_ROW_START
@@ -26,6 +30,7 @@ from curibio.sdk import NUMBER_OF_PER_TWITCH_METRICS
 from curibio.sdk import PER_TWITCH_METRICS_SHEET_NAME
 from curibio.sdk import plate_recording
 from curibio.sdk import PlateRecording
+from curibio.sdk import TWITCH_FREQUENCIES_CHART_SHEET_NAME
 from freezegun import freeze_time
 from labware_domain_models import LabwareDefinition
 from mantarray_file_manager import MAIN_FIRMWARE_VERSION_UUID
@@ -124,6 +129,28 @@ def test_write_xlsx__creates_file_at_supplied_path_with_auto_generated_name(
     pr.write_xlsx(tmp_dir)
     expected_file_name = "MA20123456__2020_08_17_145810.xlsx"
     assert os.path.exists(os.path.join(tmp_dir, expected_file_name)) is True
+
+
+def test_write_xlsx__creates_frequency_vs_time_sheet(
+    plate_recording_in_tmp_dir_for_generic_well_file_0_3_2,
+):
+    pr, tmp_dir = plate_recording_in_tmp_dir_for_generic_well_file_0_3_2
+
+    pr.write_xlsx(tmp_dir, create_waveform_charts=False)
+    expected_file_name = "MA20223322__2020_09_02_173943.xlsx"
+    actual_workbook = load_workbook(os.path.join(tmp_dir, expected_file_name))
+    assert actual_workbook.sheetnames[6] == TWITCH_FREQUENCIES_CHART_SHEET_NAME
+
+
+def test_write_xlsx__creates_force_frequency_relationship_sheet(
+    plate_recording_in_tmp_dir_for_generic_well_file_0_3_2,
+):
+    pr, tmp_dir = plate_recording_in_tmp_dir_for_generic_well_file_0_3_2
+
+    pr.write_xlsx(tmp_dir, create_waveform_charts=False)
+    expected_file_name = "MA20223322__2020_09_02_173943.xlsx"
+    actual_workbook = load_workbook(os.path.join(tmp_dir, expected_file_name))
+    assert actual_workbook.sheetnames[7] == FORCE_FREQUENCY_RELATIONSHIP_SHEET
 
 
 def test_write_xlsx__creates_per_twitch_metrics_sheet_labels(